+++ iris-mpc-common/src/iris_db/iris.rs
@@ -174,11 +174,39 @@ pub struct IrisCode {
     pub code: IrisCodeArray,
     pub mask: IrisCodeArray,
 }
+
 impl Default for IrisCode {
     fn default() -> Self {
         Self {
             code: IrisCodeArray::ZERO,
             mask: IrisCodeArray::ONES,
+        }
+    }
+}
+
+/// Iris code representation using base64 encoding compatible with Open IRIS
+#[derive(Serialize, Deserialize)]
+pub struct IrisCodeBase64 {
+    pub iris_codes: String,
+    pub mask_codes: String,
+}
+
+/// Convertor: IrisCode -> IrisCodeBase64.
+impl From<&IrisCode> for IrisCodeBase64 {
+    fn from(value: &IrisCode) -> Self {
+        Self {
+            iris_codes: value.code.to_base64().unwrap(),
+            mask_codes: value.mask.to_base64().unwrap(),
+        }
+    }
+}
+
+/// Convertor: IrisCodeBase64 -> IrisCode.
+impl From<&IrisCodeBase64> for IrisCode {
+    fn from(value: &IrisCodeBase64) -> Self {
+        Self {
+            code: IrisCodeArray::from_base64(&value.iris_codes).unwrap(),
+            mask: IrisCodeArray::from_base64(&value.mask_codes).unwrap(),
         }
     }
 }
