+++ iris-mpc-utils/Cargo.toml
@@ -54,7 +54,3 @@ path = "bin/initialise_test_dbs.rs"
 [[bin]]
 name = "write-node-config-toml"
 path = "bin/write_node_config_toml.rs"
-
-[[bin]]
-name = "write-plaintext-store"
-path = "bin/write_plaintext_store.rs"
